Casting number of internal steps to an integer

+++ dirk.py
@@ -65,7 +65,7 @@ def dirk(f, tspan, ycur, h, A, b, c, solver):
     for iout in range(1,tspan.size):
 
         # determine how many internal steps are required
-        N = round((tspan[iout]-tspan[iout-1])/h)
+        N = int(round((tspan[iout]-tspan[iout-1])/h))
 
         # reset "current" t that will be evolved internally
         tcur = tspan[iout-1]
