+++ forward_euler.py
@@ -46,7 +46,7 @@ def forward_euler(f, tspan, ycur, h):
     for iout in range(1,tspan.size):
 
         # determine how many internal steps are required
-        N = round((tspan[iout]-tspan[iout-1])/h)
+        N = int(round((tspan[iout]-tspan[iout-1])/h))
 
         # reset "current" t that will be evolved internally
         tcur = tspan[iout-1]
